p1/requirements.txt
opencv-python>=4.5.0
pytesseract>=0.3.8
Pillow>=8.0.0
numpy>=1.19.0
scikit-learn>=0.24.0
sentence-transformers>=2.0.0
transformers>=4.0.0
torch>=1.8.0
faiss-cpu>=1.7.0
openai==0.28.0
python-dotenv>=0.19.0
tqdm>=4.60.0 